requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
mongoengine==0.27.0
pymongo==4.6.1
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.70.0
